requirements.txt
fastapi
uvicorn
tensorflow
pillow
python-multipart
gradio
numpy
requests